Move test manager back into saito-core #121
moved tests
