+++ saito-core/Cargo.toml
@@ -30,3 +30,5 @@ async-trait = "0.1.52"
 
 [dev-dependencies]
 serial_test = "0.6.0"
+tokio = { version = "1.17.0", features = ["full"] }
+futures = { version = "0.3.21", default-features = false }
