+++ saito-core/src/common/mod.rs
@@ -4,3 +4,5 @@ pub mod interface_io;
 pub mod keep_time;
 pub mod process_event;
 pub mod run_task;
+pub mod test_io_handler;
+pub mod test_manager;
