+++ saito-core/src/core/data/block.rs
@@ -1905,12 +1905,14 @@ impl Block {
 
 #[cfg(test)]
 mod tests {
+    use crate::common::test_manager::test::TestManager;
     use crate::core::data::block::{Block, BlockType};
     use crate::core::data::crypto::verify;
     use crate::core::data::slip::Slip;
     use crate::core::data::transaction::{Transaction, TransactionType};
     use crate::core::data::wallet::Wallet;
     use ahash::AHashMap;
+    use futures::future::join_all;
     use hex::FromHex;
 
     #[test]
@@ -2124,5 +2126,49 @@ mod tests {
 
         assert!(block.get_merkle_root().len() == 32);
         assert_ne!(block.get_merkle_root(), [0; 32]);
+    }
+    #[tokio::test]
+    #[serial_test::serial]
+    // downgrade and upgrade a block with transactions
+    async fn block_downgrade_upgrade_test() {
+        let mut t = TestManager::new();
+        let mut wallet_lock = t.wallet_lock.clone();
+        let mut block = Block::new();
+        let mut transactions = join_all((0..5).into_iter().map(|_| async {
+            let mut transaction = Transaction::new();
+            let wallet = wallet_lock.read().await;
+            transaction.sign(wallet.get_privatekey());
+            transaction
+        }))
+        .await
+        .to_vec();
+
+        block.transactions = transactions;
+        block.generate();
+
+        // save to disk
+        t.storage.write_block_to_disk(&mut block).await;
+
+        assert_eq!(block.transactions.len(), 5);
+        assert_eq!(block.get_block_type(), BlockType::Full);
+
+        let serialized_full_block = block.serialize_for_net(BlockType::Full);
+        block
+            .update_block_to_block_type(BlockType::Pruned, &mut t.storage)
+            .await;
+
+        assert_eq!(block.transactions.len(), 0);
+        assert_eq!(block.get_block_type(), BlockType::Pruned);
+
+        block
+            .update_block_to_block_type(BlockType::Full, &mut t.storage)
+            .await;
+
+        assert_eq!(block.transactions.len(), 5);
+        assert_eq!(block.get_block_type(), BlockType::Full);
+        assert_eq!(
+            serialized_full_block,
+            block.serialize_for_net(BlockType::Full)
+        );
     }
 }
