+++ saito-rust/src/test/mod.rs
@@ -1,9 +1,1 @@
-mod block_tests;
-mod blockchain_tests;
-mod blockring_tests;
-mod mempool_tests;
-mod storage_tests;
-mod test_io_handler;
-pub mod test_manager;
 mod test_setup;
-mod wallet_tests;
